+++ R/tar_glimpse.R
@@ -26,6 +26,12 @@
 #'   is far from 1. If `level_separation` is `NULL`,
 #'   the `levelSeparation` argument of `visHierarchicalLayout()`
 #'   defaults to `150`.
+#' @param degree_from Integer of length 1. When you click on a node,
+#'   the graph highlights a neighborhood of that node. `degree_from`
+#'   controls the number of edges the neighborhood extends upstream.
+#' @param degree_to Integer of length 1. When you click on a node,
+#'   the graph highlights a neighborhood of that node. `degree_to`
+#'   controls the number of edges the neighborhood extends downstream.
 #' @examples
 #' if (identical(Sys.getenv("TAR_INTERACTIVE_EXAMPLES"), "true")) {
 #' tar_dir({ # tar_dir() runs code from a temporary directory.
@@ -47,18 +53,28 @@ tar_glimpse <- function(
   exclude = ".Random.seed",
   level_separation = NULL,
   callr_function = callr::r,
-  callr_arguments = targets::callr_args_default(callr_function)
+  callr_arguments = targets::callr_args_default(callr_function),
+  degree_from = 1L,
+  degree_to = 1L
 ) {
   assert_script()
   assert_package("visNetwork")
   assert_lgl(targets_only, "targets_only must be logical.")
+  assert_scalar(degree_from, "degree_from must have length 1.")
+  assert_scalar(degree_to, "degree_to must have length 1.")
+  assert_dbl(degree_from, "degree_from must be numeric.")
+  assert_dbl(degree_to, "degree_to must be numeric.")
+  assert_ge(degree_from, 0L, "degree_from must be at least 0.")
+  assert_ge(degree_to, 0L, "degree_to must be at least 0.")
   assert_callr_function(callr_function)
   assert_list(callr_arguments, "callr_arguments mut be a list.")
   targets_arguments <- list(
     targets_only = targets_only,
     allow_quosure = rlang::enquo(allow),
     exclude_quosure = rlang::enquo(exclude),
-    level_separation = level_separation
+    level_separation = level_separation,
+    degree_from = degree_from,
+    degree_to = degree_to
   )
   callr_outer(
     targets_function = tar_glimpse_inner,
@@ -73,7 +89,9 @@ tar_glimpse_inner <- function(
   targets_only,
   allow_quosure,
   exclude_quosure,
-  level_separation
+  level_separation,
+  degree_from,
+  degree_to
 ) {
   network <- glimpse_init(pipeline)
   visual <- visnetwork_init(
@@ -81,7 +99,9 @@ tar_glimpse_inner <- function(
     targets_only = targets_only,
     allow = allow_quosure,
     exclude = exclude_quosure,
-    level_separation = level_separation
+    level_separation = level_separation,
+    degree_from = degree_from,
+    degree_to = degree_to
   )
   visual$update()
   visual$visnetwork